+++ src/main/java/be/baur/sdt/statements/CopyStatement.java
@@ -2,12 +2,11 @@ package be.baur.sdt.statements;
 
 import java.util.List;
 
-import be.baur.sda.Node;
 import be.baur.sda.DataNode;
+import be.baur.sda.Node;
 import be.baur.sdt.StatementContext;
 import be.baur.sdt.TransformContext;
 import be.baur.sdt.TransformException;
-import be.baur.sdt.serialization.Attribute;
 import be.baur.sdt.serialization.Statements;
 import be.baur.sdt.xpath.SDAXPath;
 
@@ -73,7 +72,7 @@ public class CopyStatement extends XPathStatement {
 	@Override
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.COPY.tag);
-		node.add( new DataNode(Attribute.SELECT.tag, getExpression()) ); 
+		node.add( new DataNode(Statements.SELECT.tag, getExpression()) ); 
 		return node;
 	}