+++ src/main/java/be/baur/sdt/statements/VariableStatement.java
@@ -7,7 +7,6 @@ import be.baur.sda.DataNode;
 import be.baur.sdt.StatementContext;
 import be.baur.sdt.TransformContext;
 import be.baur.sdt.TransformException;
-import be.baur.sdt.serialization.Attribute;
 import be.baur.sdt.serialization.Statements;
 import be.baur.sdt.xpath.SDAXPath;
 
@@ -105,7 +104,7 @@ public class VariableStatement extends XPathStatement {
 	@Override
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.VARIABLE.tag, varName);
-		node.add( new DataNode(Attribute.SELECT.tag, getExpression()) ); 
+		node.add( new DataNode(Statements.SELECT.tag, getExpression()) ); 
 		return node;
 	}